% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_dir_from_path.R
\name{get_sim_dir_from_path}
\alias{get_sim_dir_from_path}
\title{Get the sim directory name from a full path}
\usage{
get_sim_dir_from_path(path, dir_split_pattern = "/",
  param_split_pattern = "_")
}
\description{
Returns just the final folder name of the simulation given its full path name.
This is really a convenience function that can then be used to parse the folder
name to get the parameteres used for a simulation
}

