% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sim_params_from_dir_name.R
\name{parse_sim_params_from_dir_name}
\alias{parse_sim_params_from_dir_name}
\title{Parse simulation parameters from a batch run}
\usage{
parse_sim_params_from_dir_name(dir_param_string, param_split = "_")
}
\description{
Multiple calls to `parse_single_param_from_dir_name`.
Usually ran on a batch/sweep MANN run.
}

