% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sim_params_from_dir_name.R
\name{parse_single_param_from_dir_name}
\alias{parse_single_param_from_dir_name}
\title{Parse parameters from a single directory path}
\usage{
parse_single_param_from_dir_name(single_param_string)
}
\description{
Takes a string that represents a parameter and value
Typically this comes from the folder name, e.g., a250 would mean 250 agents
Returns a list where the name is the string, and value is the numer
For example, 'a250' would return a list where 'a' is 250
}

